Make `buttons` only as long as it needs to be
In Draft Mode, the team builder's "Save" "Load" and "Pick Units" buttons
do not exist - they aren't even created. And yet the TeamBuilderStage
assumes in quite a few places that there are four buttons and that the
`buttons` array contains exactly four members. This change removes these
assumptions, making the `buttons` array variable sized and making the
boundary conditions act on the size of the array.