Enable jar file for rise use
Also, added rise weapon icon. This patch was a pretty easy one, so note that
FEResources not finding anything is likely the server's fault. Note that summon is the only instance
where res is necessary in FEServer, so if the filesize needs to be trimmed, that's a good place to start.

I'm hoping git remembered to check my reassignment of RiseTome's ID to 73, if not I will have to fix that later.