Allow to end turn when selecting an invalid unit
Currently, selecting (pressing 'Z') on a tile with either a friendly
unit that already moved (greyed out) or an enemy unit has no action.

This patch makes it so that instead, it will bring out the "End turn"
menu, which is the same behavior as when selecting an empty tile. This
is coherent with the behavior of the GBA games (on the GBA games,
selecting an enemy unit would display their movement and attack ranges,
but FEMP already does this when simply hovering over the unit).