Add an additional "special" key
Inspired by /u/Cardalilyn on reddit.

Currently, there are two mains keys in FEMP, 'Z' and 'X'. They
correspond to the "OK" and "Cancel" behavior of the 'A' and 'B' buttons
on the GBA games. This patch adds a third key, 'C', for implementing
"next unit" behavior. It currently does two things, taken from the
behavior of the 'L' button in the GBA games:

 - When a unit is selected, pressing 'C' will place the cursor on the
   selected unit immediately.

 - When no unit is selected, pressing 'C' will rotate the cursor through
   the player-controlled units that have not moved yet.