Remember (locally) the path
Currently, the displayed path for a unit is always a computed shortest

path. This sometimes causes unnecessary "path jumps" (for instance, if

you go one down then one left, the path may immediately change to one

left then one down).



This patch makes it so that FEMP emulates the behavior of the GBA games:

if the unit still has enough movement remaining, the existing path is

extended (and if the new tile was already on the path, it is simply

truncated). In case there is not enough movement left, the old behavior

is kept.



This is a purely local change: even though the current player will see

the unit follow the selected path, a networked opponent or spectator

will possibly see a different path that corresponds to the shortest

path. This is a compromise that allows to keep the current efficiency of

FEMP sending only the new end position (not the complete path) over the

network while providing a more user-friendly behavior inspired from the

GBA games.