Hide the path and indicators while moving
Inspired by /u/Cardalilyn on reddit.



Currently when a character moves the path arrow as well as the tile

indicators are still displayed, while on the GBA games it disappears as

soon as the destination is selected.  This patch makes it so that FEMP

behaves like the GBA games in that regard.