Move fast when pressing 'X'
Inspired by /u/Cardalilyn on reddit.



Currently, the cursor speed on the map is fixed to a relatively slow

speed suitable for selecting a unit and giving it orders. However, the

current speed is ill suited for quickly moving across the map to access

a unit there. This patch makes it so that when the 'X' button is

pressed, the cursor moves faster, emulating the behavior of the GBA

games.