Probably fixed skip turn bug
I say probably, because I can't test it on this computer - I was unable
to replicate the bug in the first place.  At any rate, players whose
turn it is not should not be able to end the turn.  I did not want to do
the same thing for CommandMessages until I am sure that there are none
that are intended to be allowed on the enemy's turn.

Because I know I'm going to get a comment about it, there are some of
the bugs with this that aren't going to be possible to fix server-side
without a moderate rewrite of the netcode, and either that or the
client-side fix will take more time than I have tonight.