Include 'team' in Unit hash
Stage::entities is a TreeSet using SortByUpdate as its ordering.
That ordering sorts by priority, then uses hashCode as a tiebreaker.
Objects aren't added to a TreeSet if they compare 0 to another object
in that TreeSet. Before this, two units of the same character but on
opposite teams would have the same hashCode. Thus, only one of them
would be added to the OverworldStage.