Assorted networking things (#102)
* Change statistics storage from a HashMap to a dedicated class

There have been far too many instances of forgetting to account for a
new statistic, especially when having to add some together. Now it is a
compile-time error instead of a run-time error.

* Realign tests

* Add StatisticsTest

* Change unit::battleStats from a HashMap to a POJO

* Remove FEMulitplayer::map.

* Make CommandMessage more structured

* Add a few automated command tests.

* Delete FEServer::maps and FEServer::getUnit

* Delete several static Game members

* * Prune RNG
* Prune FEMultiplayer
* Give all Items hashCode and equals methods

* Add equals and hashCode to Class

* Mark members as final

* Reduce scope of modifier actions

* Add a set of runtime checks validating teams on the server

* Add check for unit stats for server team validation

* Fix unchecked warnings in FEServer

* Delete unused SessionUpdate class

* Reduce session mutability

* Replace Game::messages with method parameters

* Prevent one player from ending another player's turn

* Sequentialize ClientOverworldStage message responses

* Add KeyboardEvent::toString

* Add information to ClientInit

to allow a client to reject a connection to the server.

Also, changes exist that, better than allowing connections but not
ideal, will cause the server to crash if a 1.4-or-earlier client
attempts to connect to a this server.

* Repair "Every client thinks it is id 0" issue

* Show the same chat in each window

* Allow session updates during any stage

* Reduce program exits from other clients quitting

* Ensure that ClientOverworld progresses to results on player quit

Also, *correctly* detect a disconnected nonspectator. Assuming two players.

* Prevent EndGameStage from changing due to other player's lobby actions

* Revert MadeInChina to initOverworld implementation

* Add a level hash to ClientInit.Hashes

Includes adding Level::hashCode and making Level members final

* Don't send empty messages on turn end

* Convert moving to a Command

* On a failed hash check, return to ConnectStage with notice

* restore FEMultiplayer::connect
* Make Weapon::hashCode more extensive
* Make a backwards-incompatible change to JoinLobby

* Fix a comment

* Restore ability to load teams with leveled units

* Prevent ClientOverworldStage from accepting inputs during a battle animation

FightStage and FightOverworldTransition have callbacks now to support this

* Prevent units from switching weapons mid-round

* Include 'team' in Unit hash

Stage::entities is a TreeSet using SortByUpdate as its ordering.
That ordering sorts by priority, then uses hashCode as a tiebreaker.
Objects aren't added to a TreeSet if they compare 0 to another object
in that TreeSet. Before this, two units of the same character but on
opposite teams would have the same hashCode. Thus, only one of them
would be added to the OverworldStage.

* Make Physic's range semantically Mag/2

* Allow mirror units without breaking the draft

By, instead of adding team to Unit::hashCode, instead changing
Stage::entities from a set that thinks HashCode is consistent with
equals into a list.

* Not sorting Stage entities by updatePriority

since, apparently, nothing has a non-zero updatePriority!

* Perform the re-equip after trade nonlocally

* Mark things as final for process of elimination

* Make Unit::getCopy a deeper copy

So that one copy using a weapon doesn't affect any other copies.

* Equip a usable weapon at start of AttackPreview non-locally.

* Fix concurrent modification exception.

* Disable inputs during item healthbar animation.

* Kick instead of crash on team validation failure

* Give KickMessage a toString