Start of new rules settings
Some of the framework for the new rulesets.  There is a new ruleset
called Tournament Rules now.

-Sets the minimum level for all units to 10 and the max to 20
-Adjusts exp to compensate.

The shop is next, but the way it's set up, it's going to take some time
to architect a good way to do it.  I'm probably going to work on it more
tomorrow.