Mugshot change (#112)
* Probably fixed skip turn bug

I say probably, because I can't test it on this computer - I was unable
to replicate the bug in the first place.  At any rate, players whose
turn it is not should not be able to end the turn.  I did not want to do
the same thing for CommandMessages until I am sure that there are none
that are intended to be allowed on the enemy's turn.

Because I know I'm going to get a comment about it, there are some of
the bugs with this that aren't going to be possible to fix server-side
without a moderate rewrite of the netcode, and either that or the
client-side fix will take more time than I have tonight.

* Various small QoL Tweaks

- Marth mugshot updated to the subreddit flair
- Sudden death now works as intended.  As a side note, the fix also
necessitated "unlocking" max hp, leading to the possibility of other
effects modifiying it in the future.  (Skill that reduces opponent's max
HP?)
- The village now has 10 Avo instead of 10 Def, as was intended.  When
we actually settle on how we want it balanced, we can revisit this.

* Change to the way max HP changing is handled.

See Ray's suggestion.

* Update Statistics.java

Forgot to change the comment back

* Mugshot changed

Changed Marth mugshot to another different one.

* Re-changed the Marth mugshot again

See title

* Mugshot change

-Mugshot changed back to the subreddit mugshot