Shop limit framework
So, this is the shop limiting framework.

What it doesn't do:
-Give a GUI indication.  Debugging this took longer than I wanted it to.
I'll do it over the next few days, unless someone else would rather do
it.
-Allow for classes to weapons to be limited with a limit for each
weapon.  As it stands, each weapon can be given a limit, but you can't,
say, limit a team to two "killer" weapons.  That would require more
formal weapon classes.  Not to say I don't want to, but we'll have to
figure out how to structure that.
-GUI on the server side.  It's still just in it's modifier.
-Truly faithful to the rules as prescribed on the wiki: as a
demonstration, I left each killer weapon with one buyable.

What works:
-Limits set on the rules server limit the purchases that the player can
make.
-Units that are removed from the team in blind pick "restock" the store
with the items they lost when switched out.
-Items have no limit by default.

The inventory system utilizes a Singleton pattern, which is needed
because the ShopMenu is initialized again every time the unit build
screen is opened, so the data can't be stored there.