Weapon Classes
Weapon classes work now.  They read their classes from the (currently)
last column in the weapon txt file, with the exception of
vulnerary/concoction/elixr, which have been hard-coded (for some time,
at any rate they're in the Healing class), and Rise (also hard coded, it
is not in a class).

The classes are Killer, Slayer, Reaver, Brave, Legend, Siege, and
Healing as of now.