Fixed issue in LobbyStage where game could not start with multiple spectators
Spectators are classed as a team and as such the logic here would break if there was more than one as it would think there is more than one player on a "team" and start, one work around was to make sure that the last player who joined was one who was going to be playing, this caused the iteration of in HashMap to place them first. Either way this bug did also allow multiple players to be one on team if the order of the HashMap was manipulated correctly. 

Note: I don't actually have a Java development environment so please make sure to run tests! :)