Add "Inspect Inventory" dialog/context/button for overworld stage
(issue #129)

When in the overworld stage, in the top-most context, when the cursor is
hovering over a unit with at least one item in hir inventory, on either
player's turn, the player can press "D" to display information about the
items in the hovered unit's inventory. The displayed information matches
the information that had previously been displayed in the UnitBuilderStage.
The dialog can be dismissed either by pressing the cancel button ("X") or
by pressing "D" again.