Fixed a critical bug when trying to drop/summon when there's no available tiles
Okay, so I thought that there would always be a valid tile in which the new unit could be put in, but because of the movement restriction (i.e. can't put a paladin on a mountain), it would've been possible to get stuck in an infinite loop. Besides, the original solution was kind of hacky.