Increase a unit's skillCharge when the unit is attacked
The amount is constant per hit. The amount was set based on the class's
current inate critical ratio - 1 charge for 0 inate crit, 2 charge for 10
inate crit, and 3 charge for 20 inate crit. If a killer weapon is equipped,
then charge is increased by an additional one under the same condition.

I detail this, even though I have made no attempts to balance anything and
assume the numbers will be tweaked.