Add a skill that uses skillCharge
"Rage" is a skill that will reduce skill charge by 3 and triple
damage whenever the attacker's skill charge is greater than 3.
It is completely involuntary, however. I hear the ideal is
involuntary on enemy phase an voluntary on player phase.

The rage skill's animation doesn't use the unit's critical
animations, but does use the critical hit effect, and has a
displayed name.