RejoinMessage now causes missed messages to be sent to the client
I.e. if a message was sent to the other clients while this one was reconnecting, it will properly receive those messages.
This causes the Server.broadcastMessage method to stall in case there's delay when sending the missed message for concurrency. In practice, this doesn't matter because sending a message doesn't take a lot of time and typically, the number of message to be sent will be either 0 or (rarely) 1. I don't really know how to fix that issue. The only solution I can think of is to give every ServerListener it's own message sending thread, but that seems wasteful. Either way, this was default behavior in the previous versions so whatever.